Added Load Case to Problem
