+++ src/compas_dem/analysis/lmgc90.py
@@ -271,165 +271,13 @@ def lmgc90_solve(
     results.metadata["mu"] = mu
     results.metadata["force_time"] = force_time
     results.metadata["urf_history"] = urf_history
+    results.metadata["n_steps"] = step + 1
     results.metadata["displacement_history"] = [d.tolist() for d in displacement_history]
 
     solver.name = "LMGC90"
     solver.finalize()
 
     return results
-
-
-# def _post_processing_lmgc90(solver: "Solver", model: BlockModel) -> None:
-#     """Write LMGC90 results directly onto the BlockModel graph (in-place).
-
-#     This function mutates the model. It is kept for viewer compatibility.
-#     For a standalone serializable result use :func:`_post_processing_lmgc90_results`.
-
-#     Block attributes set
-#     --------------------
-#     transformation : :class:`compas.geometry.Transformation`
-
-#     Edge attributes set
-#     -------------------
-#     contact_point, force_magnitude, force_vector, contact_polygon, gap, status,
-#     contact_data, face_contact, edge_contact, point_contact
-#     """
-#     elements = list(model.elements())
-#     contact_data = solver.get_contacts()
-#     result = solver.last_result
-#     graph = model.graph
-
-#     for i, block in enumerate(elements):
-#         pos = np.array(result.bodies[i])
-#         rot = np.array(result.body_frames[i]).reshape(3, 3)
-#         new_frame = cg.Frame(pos, rot[0, :], rot[1, :])
-#         T = cg.Transformation.from_frame_to_frame(block.init_frame, new_frame)
-#         graph.node_attribute(block.graphnode, "transformation", T)
-
-#     _per_point_keys = [
-#         "contact_points",
-#         "force_normal",
-#         "force_tangent1",
-#         "force_tangent2",
-#         "gaps",
-#         "status",
-#     ]
-#     _new_key_name = [
-#         "contact_point",
-#         "force_normal",
-#         "force_tangent1",
-#         "force_tangent2",
-#         "gap",
-#         "status",
-#     ]
-
-#     contact_groups = {}
-#     for i in range(len(solver.last_result.interaction_coords)):
-#         body_pair = tuple(sorted(b - 1 for b in solver.last_result.interaction_bodies[i]))
-#         if body_pair not in contact_groups:
-#             contact_groups[body_pair] = []
-#         contact_groups[body_pair].append(i)
-
-#     Added_Edges = 0
-#     for pair, points in contact_groups.items():
-#         u, v = pair
-
-#         if not points:
-#             continue
-
-#         if graph.has_edge((u, v)):
-#             edge = (u, v)
-#         elif graph.has_edge((v, u)):
-#             edge = (v, u)
-#         else:
-#             if not graph.has_node(u) or not graph.has_node(v):
-#                 continue
-#             graph.add_edge(u, v)
-#             edge = (u, v)
-#             Added_Edges += 1
-
-#         graph.edge_attribute(edge, "face_contact", False)
-#         graph.edge_attribute(edge, "point_contact", False)
-#         graph.edge_attribute(edge, "edge_contact", False)
-
-#         for k, name in zip(_per_point_keys, _new_key_name):
-#             graph.edge_attribute(edge, name, [contact_data[k][i] for i in points])
-#         graph.edge_attribute(
-#             edge,
-#             "force_magnitude",
-#             float(np.linalg.norm(np.sum([result.interaction_force_magnitude[p] for p in points], axis=0))),
-#         )
-#         graph.edge_attribute(edge, "force_vector", [list(result.interaction_force_global[p]) for p in points])
-#         graph.edge_attribute(edge, "force", np.sum([result.interaction_force_global[p] for p in points], axis=0).tolist())
-
-#         contact_frames = [
-#             cg.Frame(
-#                 point=cg.Point(*result.interaction_coords[p]),
-#                 xaxis=cg.Vector(*result.interaction_tangent1[p]),
-#                 yaxis=cg.Vector(*result.interaction_normals[p]),
-#             )
-#             for p in points
-#         ]
-
-#         graph.edge_attribute(edge, "contact_frame", contact_frames[0])
-
-#         polygon_pts = [result.interaction_coords[p] for p in points]
-#         contact_pts = graph.edge_attribute(edge, "contact_point")
-#         contact_frames = graph.edge_attribute(edge, "contact_frame")
-
-#         if len(polygon_pts) >= 3:
-#             graph.edge_attribute(edge, "contact_polygon", cg.Polygon(polygon_pts))
-#             graph.edge_attribute(edge, "face_contact", True)
-#             fc = FrictionContact(points=[cg.Point(*p) for p in contact_pts])
-#             lmgc_tangent = cg.Vector(*result.interaction_tangent1[points[0]])
-#             lmgc_normal = cg.Vector(*result.interaction_normals[points[0]])
-#             lmgc_tangent2 = lmgc_normal.cross(lmgc_tangent).unitized()
-#             fc._frame = cg.Frame(contact_frames.point, lmgc_tangent, lmgc_tangent2)
-
-#             for p in points:
-#                 Ft, Fn, Fs = result.interaction_rloc[p]
-#                 fc.forces.append({"c_np": max(Fn, 0), "c_nn": max(-Fn, 0), "c_u": -Ft, "c_v": -Fs})
-#             graph.edge_attribute(edge, "contact_data", fc)
-
-#         elif len(polygon_pts) == 2:
-#             print(f"Edge contact between bodies {u} and {v} with contact points {polygon_pts}. Setting edge_contact=True.")
-#             graph.edge_attribute(edge, "edge_contact", True)
-
-#             lmgc_tangent = cg.Vector(*result.interaction_tangent1[points[0]])
-#             lmgc_tangent2 = cg.Vector(*result.interaction_tangent2[points[0]])
-
-#             ec = EdgeContact(
-#                 points=[cg.Point(*p) for p in polygon_pts],
-#                 frame=cg.Frame(
-#                     cg.Line(cg.Point(*polygon_pts[0]), cg.Point(*polygon_pts[1])).midpoint,
-#                     lmgc_tangent,
-#                     lmgc_tangent2,
-#                 ),
-#             )
-#             for p in points:
-#                 Ft, Fn, Fs = result.interaction_rloc[p]
-#                 ec.forces.append({"c_np": max(Fn, 0), "c_nn": max(-Fn, 0), "c_u": -Ft, "c_v": -Fs})
-#             graph.edge_attribute(edge, "edge_contact", True)
-#             graph.edge_attribute(edge, "contact_data", ec)
-
-#         elif len(polygon_pts) == 1:
-#             graph.edge_attribute(edge, "point_contact", True)
-
-#         else:
-#             print(f"Warning: contact between bodies {u} and {v} has no contact points.")
-
-#         contact_frames = [
-#             cg.Frame(
-#                 result.interaction_coords[p],
-#                 result.interaction_tangent1[p],
-#                 result.interaction_normals[p],
-#             )
-#             for p in points
-#         ]
-#         graph.edge_attribute(edge, "contact_frames", contact_frames)
-
-#     if Added_Edges > 0:
-#         print(f"Added {Added_Edges} edges to the model graph to account for contacts without existing edges.")
 
 
 def _contact_axes(result, p: int) -> tuple:
