+++ src/compas_dem/problem/__init__.py
@@ -1,6 +1,7 @@
-from .boundary_conditions import BoundaryConditions
+from .loadcase import BoundaryConditions
+from .loadcase import LoadCase
 from .problem import Problem
 from .results import Results
 from .solvers import Solver
 
-__all__ = ["BoundaryConditions", "Problem", "Results", "Solver"]
+__all__ = ["LoadCase", "BoundaryConditions", "Problem", "Results", "Solver"]
